+++ src/main/java/com/rebuild/server/service/notification/MessageBuilder.java
@@ -47,16 +47,6 @@ public class MessageBuilder {
 	public static Message createMessage(ID toUser, String message) {
 		return new Message(null, toUser, message, null, Message.TYPE_DEFAULT);
 	}
-	
-	/**
-	 * @param toUser
-	 * @param message
-	 * @param relatedRecord
-	 * @return
-	 */
-	public static Message createMessage(ID toUser, String message, ID relatedRecord) {
-		return new Message(null, toUser, message, relatedRecord, Message.TYPE_DEFAULT);
-	}
 
 	/**
 	 * @param fromUser
@@ -67,34 +57,36 @@ public class MessageBuilder {
 	public static Message createMessage(ID fromUser, ID toUser, String message) {
 		return new Message(fromUser, toUser, message, null, Message.TYPE_DEFAULT);
 	}
-	
-	/**
-	 * @param fromUser
-	 * @param toUser
-	 * @param message
-	 * @param relatedRecord
-	 * @return
-	 */
-	public static Message createMessage(ID fromUser, ID toUser, String message, ID relatedRecord) {
-		return new Message(fromUser, toUser, message, relatedRecord, Message.TYPE_DEFAULT);
-	}
-
-	/**
-	 * @param toUser
-	 * @param message
-	 * @return
-	 */
-	public static Message createApprovalMessage(ID toUser, String message) {
-		return new Message(null, toUser, message, null, Message.TYPE_APPROVAL);
-	}
 
 	/**
 	 * @param fromUser
 	 * @param toUser
 	 * @param message
+	 * @param type
 	 * @return
 	 */
-	public static Message createApprovalMessage(ID fromUser, ID toUser, String message) {
+	public static Message createMessage(ID fromUser, ID toUser, String message, int type) {
+		return new Message(fromUser, toUser, message, null, type);
+	}
+
+	/**
+	 * @param toUser
+	 * @param message
+	 * @param recordId
+	 * @return
+	 */
+	public static Message createApproval(ID toUser, String message, ID recordId) {
+		return new Message(null, toUser, message, recordId, Message.TYPE_APPROVAL);
+	}
+
+	/**
+	 * @param fromUser
+	 * @param toUser
+	 * @param message
+	 * @param recordId
+	 * @return
+	 */
+	public static Message createApproval(ID fromUser, ID toUser, String message, ID recordId) {
 		return new Message(fromUser, toUser, message, null, Message.TYPE_APPROVAL);
 	}
