+++ src/test/java/com/rebuild/server/service/notification/MessageBuilderTest.java
@@ -16,30 +16,27 @@ You should have received a copy of the GNU General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.rebuild.server.service.notifiation;
+package com.rebuild.server.service.notification;
 
 import org.junit.Test;
 
-import com.rebuild.server.Application;
 import com.rebuild.server.TestSupport;
-import com.rebuild.server.service.notification.Message;
-import com.rebuild.server.service.notification.MessageBuilder;
+import com.rebuild.server.service.bizz.DepartmentService;
+import com.rebuild.server.service.bizz.RoleService;
+import com.rebuild.server.service.bizz.UserService;
 
 /**
  * @author devezhao-mac zhaofang123@gmail.com
  * @since 2019/03/24
  */
-public class NotificationServiceTest extends TestSupport {
+public class MessageBuilderTest extends TestSupport {
 
 	@Test
-	public void testSend() throws Exception {
-		Message minMessage = MessageBuilder.createMessage(SIMPLE_USER, "发一条消息");
-		Application.getNotifications().send(minMessage);
-		System.out.println("Notification Sent");
-	}
-	
-	@Test
-	public void testGetUnread() throws Exception {
-		Application.getNotifications().getUnreadMessage(SIMPLE_USER);
+	public void testFormatHtml() throws Exception {
+		String raw = "@" + UserService.ADMIN_USER + " 你好，这是一条消息，用户ID会特殊解析。"
+				+ "这是一条实体记录 @" + DepartmentService.ROOT_DEPT
+				+ " 可以多条，不信你看  @" + RoleService.ADMIN_ROLE;
+		String ff = MessageBuilder.toHTML(raw);
+		System.out.println("> " + raw + " \n> " + ff);
 	}
 }
